isolate ComplexEditableField
