+++ frontend/app/features/archive/components/ProductionInfoSection.tsx
@@ -1,100 +1,8 @@
-import type { Delta } from "quill";
 import { useTranslation } from "react-i18next";
-import { htmlToDelta, deltaToHtml } from "~/archive-rich-text.client";
-import { ArchiveRichTextFieldWrapper } from "~/shared/components/ArchiveRichTextFieldWrapper";
-import { Protected } from "~/features/auth";
 import { useEffect, useState } from "react";
-import { ARCHIVE_PERMISSIONS } from "../archive.constants";
-
-type Field = "teaser" | "description" | "info";
-
-type ComplexEditableFieldProps = {
-  id: string;
-  field: Field;
-  html: string | undefined;
-  fallback?: React.ReactNode;
-  isEditing: boolean;
-  onStartEdit: () => void;
-  onSave: (html: string) => void;
-  onCancel: () => void;
-  canEdit: boolean;
-};
-
-function ComplexEditableField({
-  id,
-  field,
-  html,
-  fallback,
-  isEditing,
-  onStartEdit,
-  onSave,
-  onCancel,
-  canEdit,
-}: ComplexEditableFieldProps) {
-  const { t } = useTranslation();
-  const [delta, setDelta] = useState<Delta | null>(null);
-  const shared_css = `
-    shadow-lg
-    hover:bg-archive-control-hover
-    rounded-full
-    cursor-pointer
-    transition-colors
-    duration-150
-    text-archive-ink
-    inline-flex
-    px-6 py-3
-    font-semibold text-white
-    `;
-
-  useEffect(() => {
-    if (isEditing && html) {
-      htmlToDelta(html).then(setDelta);
-    } else if (isEditing) {
-      setTimeout(() => setDelta(null), 0);
-    }
-  }, [isEditing, html]);
-
-  if (isEditing) {
-    return (
-      <Protected permissions={[ARCHIVE_PERMISSIONS.update]}>
-        <div id={id}>
-          <ArchiveRichTextFieldWrapper
-            label={field}
-            value={delta}
-            onChange={setDelta}
-            canEdit={canEdit}
-          />
-          <div className="mt-2 flex gap-2">
-            <button className={`${shared_css} bg-gray-300`} onClick={onCancel}>
-              {t("productionPage.edit.cancel")}
-            </button>
-            <button
-              className={`${shared_css} bg-archive-accent`}
-              onClick={async () => {
-                if (delta) {
-                  const html = await deltaToHtml(delta);
-                  onSave(html);
-                }
-              }}
-            >
-              {t("productionPage.edit.save")}
-            </button>
-          </div>
-        </div>
-      </Protected>
-    );
-  }
-
-  return (
-    <div
-      id={id}
-      className={`rounded opacity-90 ${canEdit ? "hover:outline-archive-accent !cursor-pointer hover:outline hover:outline-1" : "!cursor-default"}`}
-      onClick={onStartEdit}
-    >
-      {html ? <div dangerouslySetInnerHTML={{ __html: html }} /> : fallback}
-    </div>
-  );
-}
+import ComplexEditableField, {
+  type Field,
+} from "~/shared/components/ArchiveRichTextFieldWrapper";
 
 type ProductionInfoSectionProps = {
   prodinfo_available: boolean;
